// src/app/modules/academics/student-information-system/model/filter-view-model.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * 
 */

export interface FilterViewModel { 
    studentName?: Array<string>;
    admissionNumber?: Array<string>;
    feeTypeTranchId?: Array<string>;
    feeTermId?: Array<string>;
    classId?: number;
    sectionId?: number;
    facilityTypeId?: number;
    sortBy: string;
    sortOrder?: number;
    pageNumber: number;
    pageSize: number;
}