// src/app/modules/academics/student-information-system/model/student-facility-selection-model.ts
import { StudentTrancheModel } from './student-tranche-model';

/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * StudentFacilitySelection
 */
export interface StudentFacilitySelectionModel { 
    /**
     * Gets or Sets
     */
    facilityId?: number;
    /**
     * Gets or Sets
     */
    classId?: number;
    /**
     * Gets or Sets
     */
    sectionId?: number;
    /**
     * Gets or Sets
     */
    studentTrancheModel?: Array<StudentTrancheModel>;
}