// src/app/modules/academics/student-information-system/model/student-facility-assignment-edit.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * StudentFacilityResult
 */
export interface StudentFacilityAssignmentEdit { 
    /**
     * Gets or Sets FeeTypeTranche
     */
    feeTypeTrancheId?: number;
    /**
     * Gets or sets EndFeeTerm
     */
    endFeeTermId?: number;
    /**
     * Gets or Sets SectiondId
     */
    id?: string;
    /**
     * Gets or Sets FacilityTypeId
     */
    facilityTypeId?: number;
}